:hammer: Refactor EMA, SWA, & Checkpoint Ens.
