+++ experiments/classification/mnist/configs/lenet_checkpoint_ensemble.yaml
@@ -28,7 +28,7 @@ trainer:
       check_finite: true
 model:
   model:
-    class_path: torch_uncertainty.models.TrajectoryEnsemble
+    class_path: torch_uncertainty.models.CheckpointEnsemble
     init_args:
       model:
         class_path: torch_uncertainty.models.lenet._LeNet
