+++ torch_uncertainty/models/__init__.py
@@ -1,10 +1,11 @@
 # ruff: noqa: F401
 from .wrappers import (
     EMA,
+    EPOCH_UPDATE_MODEL,
+    STEP_UPDATE_MODEL,
     SWA,
+    CheckpointEnsemble,
     MCDropout,
-    TrajectoryEnsemble,
-    TrajectoryModel,
     deep_ensembles,
     mc_dropout,
 )
